figuring out downloader problems

+++ nuus/indexer.py
@@ -39,11 +39,16 @@ def enc(s):
     return s.decode('latin-1').encode('utf-8')
 
 def download_headers(id, group, start, end):
+    print '+ <%s>' % id
+    fn = os.path.join(CACHE_INBOX, CACHE_FILE_FORMAT.format(
+        group=group,page='%s-%s' % (id, start),status='new'))
+    if os.path.exists(fn):
+        print '< <%s>' % id
+        return (id, 0)
     u = usenet.Usenet(connection_pool=nuus.usenet_pool)
+    print '> <%s>' % id
     articles = u.get_articles(group, start, end)
     if articles:
-        fn = os.path.join(CACHE_INBOX, CACHE_FILE_FORMAT.format(
-            group=group,page='%s-%s' % (id, start),status='new'))
         with gzip.open(fn, 'w') as f:
             for number, subject, poster, sdate, art_id, _, size, _ in articles:
                 f.write(CACHE_LINE_FORMAT.format(
@@ -52,7 +57,11 @@ def download_headers(id, group, start, end):
                     poster=enc(poster),
                     date=parse_date(sdate),
                     size=size))
+    print '< <%s>' % id
     return (id, len(articles))
+
+def parse_header(file_name):
+    pass
 
 class Indexer(object):
     def __init__(self, articles_per_worker=10000, max_workers=8,
@@ -69,7 +78,12 @@ class Indexer(object):
         with conn.begin() as trans:
             for g in groups:
                 # get group info
+                print 'updating group:', g['group'],
                 gi = self._usenet.group_info(g['group'])
+                if g['last_post_checked']+1 < gi.last:
+                    print g['last_post_checked'], '->', gi.last
+                else:
+                    print ''
                 st = max(gi.first, g['last_post_checked']+1)
                 while st <= gi.last:
                     end = min(gi.last, st+self._articles_per_worker)
@@ -82,7 +96,7 @@ class Indexer(object):
         groups = {g['id']:g for g in groups}
         # get all tasks
         tasks = []
-        for row in conn.execute(select([tables.tasks])):
+        for row in conn.execute(select([tables.tasks]).order_by(tables.tasks.c.id)):
             tasks.append(dict(id=row['id'],
                               group=groups[row['group']]['group'],
                               start=row['start'],
@@ -96,20 +110,22 @@ class Indexer(object):
             # queue up tasks
             for task in tasks:
                 futures.append(executor.submit(download_headers, **task))
-
+            print 'all tasks queued...'
             # get results
             new_articles = 0
             tasks_complete = 0
             report_status_on = (len(tasks) / 100) or 1
             for f in as_completed(futures):
-                tid, arts = f.result()
-                conn.execute(table.tasks.delete().where(table.tasks.c.id == tid))
+                r = f.result()
+                tid, arts = r
+                r = conn.execute(tables.tasks.delete().where(tables.tasks.c.id == tid))
                 new_articles += arts
                 tasks_complete += 1
                 if tasks_complete % report_status_on == 0:
-                    print 'Completed: %s, Remaining: %s, ETR: %s' % (
+                    print '\nCompleted: %s, Remaining: %s, ETR: %s' % (
                         tasks_complete, len(tasks) - tasks_complete, 
                         time_remaining(start_time, tasks_complete, len(tasks) - tasks_complete))
+                print '- <%s>' % tid
 
 if __name__ == '__main__':
     indexer = Indexer()
@@ -118,4 +134,3 @@ if __name__ == '__main__':
         indexer.run()
     except KeyboardInterrupt:
         print 'forcing shutdown...'
-        indexer.shutdown()
